Change Naming Convention
Everything that can be processed by GNU Stow's --dotfiles mode has been
renamed s.t. it works with it. This means converting '.zshrc' ->
'dot-zshrc' and so on and so forth. The benefit of this is that my
dotfiles directory (the one I commit to VCS) is no longer filled with
hidden files. While the symlinks that live in my actual $HOME directory
are unchanged.

Nvim directory named changed to Vim, to reflect the fact that config
files for both Vim and Neovim live in the same spot. (init.vim sources
vimrc)