Change vim settings for c programming
Use 80char soft line limit for source code files, as well as setting SW
to 8.

Remove cursorline, and mouse. "set mouse=a" was really only their for
pair-ing, and I can just turn it on in that case, and this way the
cursor doesn't jump when I click the window to focus it.